Create the minimal structure for the archiving scheme.
The archiving scheme is based on the PyTangoArchiving library
and allows to access to the archiving data using taurus models.

Authority: Is the Tango database that define the archiving properties.
Device: Is a PyTangoArchiving Reader for a given archiving scheme.
Attribute: Is an archived Tango attribute.

Implement the archiving validators with its test suite.
Minimal functional implementation for Archiving Factory,
Device and Attribute.